Add files via upload
Registration.java: Has been changed to pass an intent meaning we are not just calling the .xml file now. This means we could reduce the number of objects being added to the database.
Registration_CardDetails.java: Has been changed to take in the bundle from Registration.java, unwrap it, create a customer object and add that object to the database.
Customer.Java: this is an Object class for our customer containing all the details that we take in from the sign-up form in the application.